requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv 
pymysql 
flask-login
cryptography
requests
flask-mail 
itsdangerous

bleach
email-validator
reportlab
flask-jwt-extended 
python-dotenv

# Autenticación con Google
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
authlib
